+++ src/api/objects/tag.ts
@@ -3,8 +3,9 @@
  * API docs: https://docs.joinmastodon.org/entities/Tag/
  */
 import MastoApi from "../../api/api";
-import { MastodonTag, StringNumberDict, TagNames, TagWithUsageCounts } from "../../types";
-import { removeDiacritics, wordRegex } from "../../helpers/string_helpers";
+import { detectHashtagLanguage } from "../../helpers/language_helper";
+import { MastodonTag, TagNames, TagWithUsageCounts } from "../../types";
+import { removeDiacritics } from "../../helpers/string_helpers";
 
 const BROKEN_TAG = "<<BROKEN_TAG>>";
 
@@ -12,16 +13,14 @@ const BROKEN_TAG = "<<BROKEN_TAG>>";
 // Build a lookup table of tag names to tag objects
 export function buildTagNames(tags: MastodonTag[]): TagNames {
     return tags.reduce((tagNames: TagNames, tag: MastodonTag) => {
-        const newTag = tag as TagWithUsageCounts;
-        newTag.regex ||= wordRegex(tag.name);
-        tagNames[tag.name] = newTag;
+        tagNames[tag.name] as TagWithUsageCounts;
         return tagNames;
     }, {} as TagNames);
 };
 
 
 // Lowercase the tag name, replace URL with one on homeserver
-export function repairTag(tag: MastodonTag): MastodonTag {
+export function repairTag(tag: TagWithUsageCounts): TagWithUsageCounts {
     if (!tag.name?.length) {
         console.warn(`Broken tag object:`, tag);
         tag.name = BROKEN_TAG;
@@ -36,8 +35,10 @@ export function repairTag(tag: MastodonTag): MastodonTag {
         tag.url = tag.url.toLowerCase() || "";
     }
 
+    const language = detectHashtagLanguage(tag.name);
+    if (language) tag.language = language;  // Don't set unnecessarily for storage space reasons
     return tag;
 };
 
 
-export const tagStr = (tag: TagWithUsageCounts) => `${tag.name} (${tag.numToots} numToots)`;
+export const tagStr = (tag: TagWithUsageCounts) => `${tag.name} (${tag.numToots} numToots, language: ${tag.language})`;