+++ src/api/objects/toot.ts
@@ -149,7 +149,7 @@ export default class Toot implements TootObj {
     repliesCount!: number;
     sensitive!: boolean;
     spoilerText!: string;
-    tags!: mastodon.v1.Tag[];
+    tags!: TagWithUsageCounts[];
     visibility!: mastodon.v1.StatusVisibility;
     // Optional fields
     bookmarked?: boolean | null;