+++ src/helpers/language_helper.ts
@@ -213,6 +213,8 @@ const LANGUAGE_REGEXES = {
     [LANGUAGE_CODES.arabic]: new RegExp(`^[\\p{Script=Arabic}\\d]+$`, 'v'),
     [LANGUAGE_CODES.greek]: new RegExp(`^[\\p{Script=Greek}\\d]+$`, 'v'), // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp/unicodeSets
     // TODO: this doesn't seem to match the "de" (で) character in "これを見た人は無言で"??
+    // TODO: also doesn't match half width "ga" (が) character in "やったことある人がいたら嬉しいゲーム";
+    // Both of these are weird - in some editors you can delete just the accent mark
     [LANGUAGE_CODES.japanese]: new RegExp(`^[ー・\\p{Script=Han}\\p{Script=Hiragana}\\p{Script=Katakana}]{2,}[ー・\\p{Script=Han}\\p{Script=Hiragana}\\p{Script=Katakana}\\da-z]*$`, 'iv'),
     [LANGUAGE_CODES.korean]: new RegExp(`^[\\p{Script=Hangul}\\d]+$`, 'v'), // [KOREAN_LANGUAGE]: /^[가-힣]{2,}/,
     [LANGUAGE_CODES.russian]: new RegExp(`^[\\p{Script=Cyrillic}\\d]+$`, 'v'),
@@ -333,7 +335,8 @@ export function detectLanguage(text: string): LanguageDetectInfo {
 };
 
 
-// Returns the language code of the matched regex (if any). This is our janky version of language detection.
+// Returns the language code of the matched regex (if any). Not as thorough as detectLanguage() and only
+// meant for non Latin scripts like japanese, korean, etc.
 export function detectHashtagLanguage(str: string): string | undefined {
     let language: string | undefined;